// Frontend(yeni)/Frontend/src/app/model/userInformation.ts
export class UserInformationDTO{


    id?: string;
    roleId?:string;
    //appUserId?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
    phoneNumber?: string;
    roleName?:string;
    expenseCategoryId?: string;
    expenseCategoryName?: string;
}